requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
autopep8==1.4.4
beautifulsoup4==4.8.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.7
cssselect==1.1.0
cupshelpers==1.0
evdev==1.1.2
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
jieba3k==0.35.1
lutris==0.5.3
lxml==4.4.1
mutagen==1.41.0
newspaper3k==0.2.8
nltk==3.4.5
onboard==1.4.1
packaging==19.2
Pillow==6.0.0
pyasn1==0.4.4
pycairo==1.18.0
pycodestyle==2.5.0
pycparser==2.19
pycups==1.9.74
PyGObject==3.34.0
pyOpenSSL==19.0.0
pyparsing==2.3.1
pysmbc==1.0.15.8
PySocks==1.6.7
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.21.0
requests-file==1.4.3
six==1.12.0
soupsieve==1.9.4
tinysegmenter==0.3
tldextract==2.2.2
urllib3==1.22
youtube-dl==2019.10.16
